Add calibrate function
The calibrate function os the top level controller for the automation process.
This function takes all of the parameters used in launching transCAD, setting
up and launching the abm and updating the calibration files. Depending on the
step the function will wait until a specific file is updated, kill transCAD
and all child processes and update calibration files. This will greatly increase
the speed of calibration activities and remove the human element (potential error).